Add unit tests
This adds basic unit tests for existing SQL task actions. This is in
preparation for changes to the configuration format to allow actions
to be used more flexibly (i.e. using the same action multiple times
within the same task, reordering actions). Since this will require major
changes, we want to make sure to not break existing functionality.