Fix task execution causing issues in (nested) transactions
This fixes an issue where SQL Tasks that are executed in specific
transaction scenarios (e.g. as a CiviRule action) may cause
deadlocks if further transactions are started as part of the
task execution (e.g. in API calls). The issue is fixed by
forcing a commit of all pending transactions.