GP-26375 Fix email issue when sending to multiple recipients
This fixes an issue with newer versions of Civi when sending emails
to multiple recipients using a comma-separated list of email addresses.

MessageTemplate.send no longer accepts multiple recipients in a single
call, so we call the API multiple times instead.

Shared email sending code was extracted into a trait to DRY things
up a bit.